requirements.txt
Flask
flask-cors
face-recognition
opencv-python<4.10
numpy<2
Pillow
